feat(httpapi): Add new POST /$reset endpoint
New POST /$reset endpoint. If you send a POST request to the $reset
endpoint it will reset all of the logged calls and matchers. This is
useful if you want to reset the calls & matchers in between requests
from a different process/container/language than the one that created
the sham.
- Doesn't require a body
- Returns 204 No Content

Closes #24